[GONG-38] refactor: 반 생성 단계 별로 폼 구분
