+++ src/app/class/create/step/_forms/TemplateForm.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { Button } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
 import Image from 'next/image';
 import { useState } from 'react';
@@ -12,9 +13,39 @@ export type TemplateData = {
   image: string;
 };
 
-export const TemplateList = ({ list }: { list: TemplateData[] }) => {
+type TemplateFormProp = {
+  list: TemplateData[];
+  onSuccess: (selectedId: number) => void;
+};
+
+export default function TemplateForm({ list, onSuccess }: TemplateFormProp) {
   const [selectedId, setSelectedId] = useState(0);
 
+  return (
+    <section className="w-full flex-1 flex flex-col gap-3">
+      <div>
+        <h4 className="font-bold mb-3">템플릿 목록</h4>
+        <TemplateList list={list} selectedId={selectedId} onSelect={(id) => setSelectedId(id)} />
+      </div>
+
+      <div className="flex-1 flex flex-col justify-end">
+        <Button type="button" className="w-full rounded-full" onClick={() => onSuccess(selectedId)}>
+          다음
+        </Button>
+      </div>
+    </section>
+  );
+}
+
+const TemplateList = ({
+  selectedId,
+  onSelect,
+  list,
+}: {
+  list: TemplateData[];
+  selectedId: number;
+  onSelect: (selectedId: number) => void;
+}) => {
   return (
     <ul className="flex flex-col gap-3">
       {list.map((templateData) => (
@@ -22,7 +53,7 @@ export const TemplateList = ({ list }: { list: TemplateData[] }) => {
           key={templateData.id}
           {...templateData}
           isSelected={templateData.id === selectedId}
-          onSelect={(id) => setSelectedId(id)}
+          onSelect={onSelect}
         />
       ))}
     </ul>
